stock-tracker/worker/requirements.txt
yfinance==0.2.55
pandas==2.2.3
requests==2.32.3
supabase==2.10.0
python-dotenv==1.0.1
pytz==2024.2
